backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
sentence-transformers==2.2.2
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3
torch==2.1.1
transformers==4.35.2
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0